requirements.txt
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.3
choreographer==1.0.9
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
dash==3.2.0
Flask==3.1.1
fonttools==4.59.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
kaleido==1.0.0
kiwisolver==1.4.9
logistro==1.1.0
MarkupSafe==3.0.2
matplotlib==3.10.5
narwhals==2.1.1
nest-asyncio==1.6.0
numpy==2.3.2
orjson==3.11.2
packaging==25.0
pandas==2.3.1
pillow==11.3.0
plotly==6.3.0
psycopg2-binary==2.9.9
pydantic==2.9.2
sqlalchemy==2.0.36
pydantic-settings==2.6.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
reportlab==4.4.3
requests==2.32.4
retrying==1.4.2
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
zipp==3.23.0
